change long

+++ Patch_test/test.cpp
@@ -2,7 +2,7 @@
 
 int main(int argc, char* argv[])
 {
-	int a = 10;
+	long a = 10;
 	printf("a=%d\n", a);
 	return 0;
 }
